+++ qiskit_ibm_runtime/executor.py
@@ -20,7 +20,7 @@ from typing import Any, TYPE_CHECKING
 
 from qiskit_ibm_runtime.base_primitive import get_mode_service_backend
 from qiskit_ibm_runtime.fake_provider.local_service import QiskitRuntimeLocalService
-from .options_v3.executor_options import ExecutorOptions
+from .options_models.executor_options import ExecutorOptions
 from .quantum_program.result_decoders import QuantumProgramResultDecoder
 from .quantum_program.params_converters import QUANTUM_PROGRAM_PARAMS_CONVERTERS
 from .utils.default_session import get_cm_session
