+++ qiskit_ibm_runtime/noise_learner_v3/params_converters.py
@@ -32,7 +32,7 @@ if TYPE_CHECKING:
     from ibm_quantum_schemas.common import BaseParamsModel
     from qiskit.circuit import CircuitInstruction
 
-    from qiskit_ibm_runtime.options_v3 import NoiseLearnerV3Options
+    from qiskit_ibm_runtime.options_models import NoiseLearnerV3Options
 
 
 class ParamsConverter(NamedTuple):
