+++ qiskit_ibm_runtime/quantum_program/converters/converters_0_1.py
@@ -38,7 +38,7 @@ from ibm_quantum_schemas.common import (
 
 from ..quantum_program import QuantumProgram, CircuitItem, SamplexItem
 from ..quantum_program_result import QuantumProgramResult, ChunkPart, ChunkSpan, Metadata
-from ...options_v3.executor_options import ExecutorOptions
+from ...options_models.executor_options import ExecutorOptions
 from ...utils.utils import get_qpy_version, get_ssv_version
