+++ test/unit/noise_learner_v3/converters/test_converters_0_2.py
@@ -27,7 +27,7 @@ from qiskit_ibm_runtime.noise_learner_v3.noise_learner_v3_result import (  # typ
     NoiseLearnerV3Result,
     NoiseLearnerV3Results,
 )
-from qiskit_ibm_runtime.options_v3 import NoiseLearnerV3Options
+from qiskit_ibm_runtime.options_models import NoiseLearnerV3Options
 
 from ....ibm_test_case import IBMTestCase
