+++ test/unit/noise_learner_v3/test_noise_learner_v3.py
@@ -19,7 +19,7 @@ from pydantic import ValidationError
 from test.utils import get_mocked_backend, get_mocked_session
 
 from qiskit_ibm_runtime.noise_learner_v3 import NoiseLearnerV3
-from qiskit_ibm_runtime.options_v3 import (
+from qiskit_ibm_runtime.options_models import (
     NoiseLearnerV3Options,
     PostSelectionOptions,
     ExecutionOptions,
