+++ test/unit/noise_learner_v3/test_validation.py
@@ -17,7 +17,7 @@ from qiskit import QuantumCircuit
 from samplomatic import Twirl
 
 from qiskit_ibm_runtime.noise_learner_v3.validation import validate_options, validate_instruction
-from qiskit_ibm_runtime.options_v3 import NoiseLearnerV3Options
+from qiskit_ibm_runtime.options_models import NoiseLearnerV3Options
 from qiskit_ibm_runtime.fake_provider.backends import FakeAlgiers, FakeFractionalBackend
 from qiskit_ibm_runtime.exceptions import IBMInputValueError
