+++ test/unit/quantum_program/converters/test_converters_0_2.py
@@ -39,7 +39,7 @@ from qiskit_ibm_runtime.quantum_program.converters import (
     quantum_program_result_from_0_2,
     quantum_program_from_0_2,
 )
-from qiskit_ibm_runtime.options_v3.executor_options import ExecutorOptions, ExecutionOptions
+from qiskit_ibm_runtime.options_models.executor_options import ExecutorOptions, ExecutionOptions
 
 from ....ibm_test_case import IBMTestCase
