+++ test/unit/quantum_program/converters/test_converters_1_0.py
@@ -39,7 +39,7 @@ from qiskit_ibm_runtime.quantum_program.converters import (
     quantum_program_result_from_1_0,
     quantum_program_from_1_0,
 )
-from qiskit_ibm_runtime.options_v3.executor_options import ExecutorOptions, ExecutionOptions
+from qiskit_ibm_runtime.options_models.executor_options import ExecutorOptions, ExecutionOptions
 
 from ....ibm_test_case import IBMTestCase
